v2.0 rev4335 (16 Oct 2023)
- Now requires Java 8 or higher
- Added support for ending video of Aconcagua - thanks to Martin Korth (no$psx developer)
- Added Aconcagua frame replacing support
- Optimized player and fixed minor memory leak
- Improved error detection and handling
- Added reference sources, specifications and documents
Known functional issues
- Translations (Spanish, Italian, machine translated Japanese) not updated since v1.00
- Frame rate detection still not as good as I want
- Real-time playback doesn't work for videos with a long audio delay
- Real-time playback stops early if audio ends before video ends
- Several games still not supported (see Github issues list)

+++ jpsxdec/src/jpsxdec/psxvideo/mdec/idct/PsxMdecIDCT_int.java
@@ -1,6 +1,6 @@
 /*
  * jPSXdec: PlayStation 1 Media Decoder/Converter in Java
- * Copyright (C) 2007-2020  Michael Sabin
+ * Copyright (C) 2007-2023  Michael Sabin
  * All rights reserved.
  *
  * Redistribution and use of the jPSXdec code or any derivative works are
@@ -75,7 +75,7 @@ public class PsxMdecIDCT_int implements IDCT_int {
                 tempSum = 0;
 
                 for (i=0; i<8; i++) {
-                    tempSum += idctMatrix[x + i*8] * PSX_DEFAULT_COSINE_MATRIX[y + i*8];
+                    tempSum += (long)idctMatrix[x + i*8] * PSX_DEFAULT_COSINE_MATRIX[y + i*8];
                 }
 
                 _aTemp[x + y*8] = tempSum;
